simple fix telegrambot api 1

+++ src/main/kotlin/com/github/kotlintelegrambot/entities/payments/LabeledPrice.kt
@@ -1,8 +1,7 @@
 package com.github.kotlintelegrambot.entities.payments
 
-import java.math.BigInteger
 
 data class LabeledPrice(
     val label: String? = null,
-    val amount: BigInteger,
+    val amount: Long,
 )
